requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.2
django-cleanup==8.1.0
django-filter==24.3
django-ipware==7.0.1
django-jazzmin==3.0.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-role-permissions==3.2.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.8
jsonfield==3.1.0
outcome==1.3.0.post0
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
python-ipware==3.0.0
pytz==2024.2
requests==2.32.3
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
swapper==1.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0